Serve discovered pods as clusters
This commit adds an XDS server and returns discovered pods as clusters. The design, at the moment,
is very simplistic. It will write every cluster to the cache for every registered node on every
watch event.

I've also added a sample envoy.yaml under sample/ and a Makefile to run it with docker.